requirements.txt
python-telegram-bot==20.8
requests==2.31.0
geopy==2.4.1
python-dotenv==1.0.0 